Add amdgcn machine directory (AMDGPU-22).
This patch adds an amdgcn machine directory to newlib, and the
configure support to build it, now that building newlib no longer ICEs
but runs into the absence of such a directory.

The directory is based on that for nvptx, but with most of the sources
needed; a minimal substantive port of at least exit and I/O support
may be needed, but that's a separate issue.  The nvptx version of
abort was kept, since calling __builtin_trap should be equally
applicable for GCN (however, this version currently fails to build
because of an assembler error, AMDGPU-88).